Main commit
A generic data type for a deque and a randomized queue. The goal of this assignment is to implement elementary data structures using arrays and linked lists, generics and iterators.
 A randomized queue is similar to a stack or queue, except that the item removed is chosen uniformly at random among items in the data structure.
The running time of Permutation is linear in the size of the input. Use only a constant amount of memory plus either one Deque or RandomizedQueue object of maximum size at most n.
Course:
https://www.coursera.org/learn/algorithms-part1/programming/zamjZ/deques-and-randomized-queues

Libraries:
1)))))https://algs4.cs.princeton.edu/code/javadoc/edu/princeton/cs/algs4/StdRandom.html
2))))))https://algs4.cs.princeton.edu/code/javadoc/edu/princeton/cs/algs4/StdIn.html
3))))))https://algs4.cs.princeton.edu/code/javadoc/edu/princeton/cs/algs4/StdOut.html